Break out resolvers/details into separate files
