+++ lib/device_wizard.rb
@@ -1,18 +1,11 @@
-require 'device_wizard/android_resolver'
-require 'device_wizard/browser_details'
-require 'device_wizard/device_details'
+# frozen_string_literal: true
+
+require 'device_wizard/details'
 require 'device_wizard/device_type'
-require 'device_wizard/firefox_resolver'
-require 'device_wizard/google_chrome_resolver'
-require 'device_wizard/internet_explorer_resolver'
-require 'device_wizard/ios_resolver'
-require 'device_wizard/mac_resolver'
-require 'device_wizard/operating_system_details'
-require 'device_wizard/safari_resolver'
+require 'device_wizard/resolvers'
 require 'device_wizard/user_agent_detector'
 require 'device_wizard/version'
-require 'device_wizard/windows_resolver'
 
 module DeviceWizard
-  UNKNOWN = 'Unknown'.freeze
+  UNKNOWN = 'Unknown'
 end
