+++ lib/device_wizard/details.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+require 'device_wizard/details/browser'
+require 'device_wizard/details/device'
+require 'device_wizard/details/operating_system'
