+++ lib/device_wizard/details/device.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+module DeviceWizard
+  module Details
+    class Device
+      attr_accessor :type
+      attr_accessor :browser
+      attr_accessor :os
+    end
+  end
+end
