+++ lib/device_wizard/details/operating_system.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+module DeviceWizard
+  module Details
+    class OperatingSystem
+      attr_accessor :name
+      attr_accessor :type
+      attr_accessor :version
+    end
+  end
+end
