--- lib/device_wizard/device_details.rb
@@ -1,7 +1,0 @@
-module DeviceWizard
-  class DeviceDetails
-    attr_accessor :type
-    attr_accessor :browser
-    attr_accessor :os
-  end
-end
