--- lib/device_wizard/operating_system_details.rb
@@ -1,7 +1,0 @@
-module DeviceWizard
-  class OperatingSystemDetails
-    attr_accessor :name
-    attr_accessor :type
-    attr_accessor :version
-  end
-end
