+++ lib/device_wizard/resolvers.rb
@@ -1,0 +1,12 @@
+# frozen_string_literal: true
+
+require 'device_wizard/resolvers/base'
+
+require 'device_wizard/resolvers/android'
+require 'device_wizard/resolvers/firefox'
+require 'device_wizard/resolvers/google_chrome'
+require 'device_wizard/resolvers/internet_explorer'
+require 'device_wizard/resolvers/ios'
+require 'device_wizard/resolvers/mac'
+require 'device_wizard/resolvers/safari'
+require 'device_wizard/resolvers/windows'
