+++ lib/device_wizard/resolvers/mac.rb
@@ -1,0 +1,20 @@
+# frozen_string_literal: true
+
+module DeviceWizard
+  module Resolvers
+    class Mac < Base
+      NAME = 'Mac OSX'
+      KEYWORD = 'intel mac os x'
+      REGEX = Regexp.new('intel mac os x ([0-9]+_[0-9])')
+
+      def get_version(user_agent)
+        result = super(user_agent)
+        result.gsub('_','.')
+      end
+
+      def details_klass
+        Details::OperatingSystem
+      end
+    end
+  end
+end
