+++ lib/device_wizard/version.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 module DeviceWizard
   VERSION = "0.3.0"
 end
